requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
sqlparse==0.5.3
tzdata==2025.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1